+++ examples/kyb-app/src/hooks/useSignin/types.ts
@@ -1,3 +1,0 @@
-export interface SigninCredentials {
-  email: string;
-}