+++ services/workflows-service/src/collection-flow/collection-flow.module.ts
@@ -1,7 +1,7 @@
 import { PasswordService } from '@/auth/password/password.service';
 import { BusinessRepository } from '@/business/business.repository';
 import { BusinessService } from '@/business/business.service';
-import { ColectionFlowController } from '@/collection-flow/collection-flow.controller';
+import { ColectionFlowController } from '@/collection-flow/controllers/collection-flow.controller';
 import { CollectionFlowService } from '@/collection-flow/collection-flow.service';
 import { WorkflowAdapterManager } from '@/collection-flow/workflow-adapter.manager';
 import { AppLoggerModule } from '@/common/app-logger/app-logger.module';
@@ -27,12 +27,21 @@ import { WorkflowService } from '@/workflow/workflow.service';
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { CustomerModule } from '@/customer/customer.module';
+import { TokenAuthModule } from '@/common/guards/token-guard/token-auth.module';
+import { CustomerService } from '@/customer/customer.service';
+import { CustomerRepository } from '@/customer/customer.repository';
+import { CollectionFlowFilesController } from '@/collection-flow/controllers/collection-flow.files.controller';
+import { CollectionFlowBusinessController } from '@/collection-flow/controllers/collection-flow.business.controller';
 import { SalesforceService } from '@/salesforce/salesforce.service';
 import { SalesforceIntegrationRepository } from '@/salesforce/salesforce-integration.repository';
 
 @Module({
-  imports: [AppLoggerModule, HttpModule, ProjectModule, CustomerModule],
-  controllers: [ColectionFlowController],
+  imports: [AppLoggerModule, HttpModule, ProjectModule, CustomerModule, TokenAuthModule],
+  controllers: [
+    ColectionFlowController,
+    CollectionFlowFilesController,
+    CollectionFlowBusinessController,
+  ],
   providers: [
     CollectionFlowService,
     EndUserService,
@@ -57,6 +66,10 @@ import { SalesforceIntegrationRepository } from '@/salesforce/salesforce-integra
     UserService,
     UserRepository,
     PasswordService,
+    CustomerService,
+    CustomerRepository,
+    StorageService,
+    FileRepository,
     SalesforceService,
     SalesforceIntegrationRepository,
   ],