+++ services/workflows-service/src/collection-flow/workflow-adapters/kyb_parent_kyc_session_example/kyb_parent_kyc_session_example.workflow-adapter.ts
@@ -49,12 +49,12 @@ export class KYBParentKYCSessionExampleAdapter
             entity: {
               id: ubo.id,
               type: 'individual',
+
               data: {
                 firstName: ubo.firstName,
                 lastName: ubo.lastName,
                 email: ubo.email,
                 dateOfBirth: ubo.birthDate,
-
                 additionalInfo: {
                   normalizedCustomerCompany: customer.name,
                   companyName: payload.businessData.companyName || '',
@@ -86,8 +86,6 @@ export class KYBParentKYCSessionExampleAdapter
         pages: [
           {
             ballerineFileId: document.fileId,
-            uri: document.uri,
-            provider: document.uri ? 'http' : undefined,
           },
         ],
         properties: document.properties,