+++ packages/common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/common
 
+## 0.7.10
+
+### Patch Changes
+
+- efedc9cb: Update document schemas
+- Version bump
+
 ## 0.7.5
 
 ### Patch Changes